requirements.txt
autopep8==1.4.3
coverage==4.5.2
Django==2.0.7
django-crispy-forms==1.7.2
django-filter==2.1.0
django-simple-email-confirmation==0.23
django-six==1.0.4
djangorestframework==3.9.1
entrypoints==0.3
flake8==3.7.7
isort==4.3.10
Markdown==3.0.1
mccabe==0.6.1
mysqlclient==1.4.1
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2018.9